Fingerprint all images to allow aggressive caching
Previously images loaded at URLs like:

```
/how-to-undervolt-gpu/kill-a-watt.png
```

Now the image hash is included in the URL

```
/how-to-undervolt-gpu/kill-a-watt.946b15cf74a370d49f99a057846a9788ca818990dbc21d3d678b4f3e26300a5e.png
```

The upside is that extremely aggressive caching can be used, as if the
image changed then it's hash (and thus URL) will change.

The downside is that this won't allow others to hotlink images. I'm ok
with this tradeoff as either the images should be rehosted or the
article directly linked.

The other downside is that if someone is reading an article and the site
is updated then images that have not yet loaded will 404 out as the HTML
could be referencing deleted images. I think this would be a rare
occurrence as I'd have to update an image as a user is starting to read
an article (though it is possible someone could land on the page, store
in a tab for 6 weeks, and then come back and start scrolling).